moves scale score service to package that uses it (assessments module)
