+++ webapp/src/main/webapp/src/app/shared/authentication/authentication.service.ts
@@ -1,7 +1,6 @@
 import { Injectable } from "@angular/core";
 import { Router } from "@angular/router";
-import { StorageService, StorageType } from "../../../../../../../../common-ngx/src/core/storage.service";
-import { WindowRefService } from "../window-ref.service";
+import { StorageService, StorageType, WindowRefService } from "@sbac/rdw-reporting-common-ngx";
 
 /**
  * This service is responsible for handling authentication errors.
